+++ api/views.py
@@ -1,51 +1,2 @@
 from django.shortcuts import render
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from .models import *
-from .serializers import *
 
-
-class UserApiView(APIView):
-
-    def get(self, request):
-        user_data = User.objects.all()
-        return Response({'users': UserSerializer(user_data, many=True).data})
-
-    def post(self, request):
-        serializer = UserSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        new_user = User.objects.create(
-            name=request.data['name'],
-            email=request.data['email'],
-            password=request.data['password']
-        )
-        return Response({'post': UserSerializer(new_user).data})
-
-    def put(self, request, *args, **kwargs):
-        pk = kwargs.get("pk", None)
-        if not pk:
-            return Response({"error": "Method PUT not allowed"})
-
-        try:
-            instance = User.objects.get(pk=pk)
-        except:
-            return Response({"error": "Object does not exists"})
-
-        serializer = UserSerializer(data=request.data, instance=instance)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response({"post": serializer.data})
-
-    def delete(self, request, *args, **kwargs):
-        pk = kwargs.get("pk", None)
-        if not pk:
-            return Response({"error": "Method DELETE not allowed"})
-
-        try:
-            instance = User.objects.get(pk=pk)
-        except:
-            return Response({"error": "Object does not exists"})
-
-        instance.delete()
-        return Response({"user": "delete user " + str(pk)})
